+++ changelog.md
@@ -2,7 +2,17 @@
 
 All notable changes to this project will be documented in this file.
 
-The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
+
+## [2.0.0]—2026.02.01
+
+### Added
+
+Module setting to control whether to toggles minimization or make no changes to open windows.
+
+### Changed
+
+- Compatible with Foundry 13
 
 ## [1.0.0]​—2022.21.01
 
@@ -19,4 +29,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Changed
 
 - Fixed the minimization of non-minimizable windows. Thanks to [@flamewave000](https://github.com/flamewave000).
+
+
+